Feature/abstract datasets and ids types

+++ backend/core/app/models/dataset.py
@@ -1,41 +1,46 @@
-from sqlalchemy import Column, Integer, String
-from sqlalchemy.types import BLOB
-from sqlalchemy.orm import Session, defer
+from sqlalchemy import Column, Integer, String, ForeignKey
+from sqlalchemy.orm import Session, relationship, selectinload
+from ..database import Base, get_db_session_context
+from sqlalchemy.future import select
+from sqlalchemy.ext.asyncio import AsyncSession
 import aiofiles
-from ..database import Base
 
 class Dataset(Base):
     __tablename__ = "dataset"
 
     id = Column(Integer, primary_key=True, autoincrement=True)
     name = Column(String(64), nullable=False)
-    pcap_file_path = Column(String(1024), nullable=False)
+    data_file_path = Column(String(1024), nullable=False)
     labels_file_path = Column(String(1024), nullable=False)
     description = Column(String(2048), nullable=False)
     ammount_benign = Column(Integer, nullable=False)
     ammount_malicious = Column(Integer, nullable=False)
+    dataset_type_id = Column(Integer, ForeignKey("dataset_type.id"), nullable=False)
 
-    async def read_pcap_file(self):
-        async with aiofiles.open(self.pcap_file_path, 'rb') as file:
-            pcap_file = await file.read()
-            return pcap_file
+    dataset_type = relationship('DatasetType', back_populates="dataset", lazy="selectin")
 
 
-def get_dataset_by_id(db: Session, dataset_id: int):
-    return db.query(Dataset).filter(Dataset.id == dataset_id).first()
-    
+async def get_dataset_by_id(db: AsyncSession, dataset_id: int):
+    stmt = select(Dataset).where(Dataset.id == dataset_id)
+    result = await db.execute(stmt)
+    return result.scalar_one_or_none() 
 
-def get_all_datasets(db: Session):
-    return db.query(Dataset).all()
+async def get_all_datasets(db: AsyncSession):
+    stmt = select(Dataset)
+    result = await db.execute(stmt)
+    return result.scalars().all()  
 
-def remove_dataset_by_id(db: Session, id: int):
+async def remove_dataset_by_id(db: AsyncSession, id: int):
     from ..utils import remove_directory
-    dataset: Dataset = get_dataset_by_id(db, id)
-    directory = "/".join(dataset.labels_file_path.split("/")[:-2])
-    remove_directory(directory)
-    db.delete(dataset)
-    db.commit()
-
-def add_dataset(db: Session, dataset: Dataset):
-    db.add(dataset)
-    db.commit()
+    dataset: Dataset = await get_dataset_by_id(db, id)
+    if dataset:
+        print(dataset.labels_file_path)
+        directory = "/".join(dataset.labels_file_path.split("/")[:-1])
+        print(directory)
+        remove_directory(directory)
+        await db.delete(dataset)
+        await db.commit()
+async def add_dataset(db: AsyncSession, dataset: Dataset):
+        db.add(dataset)
+        await db.commit() 
+        await db.refresh(dataset)
